services/ai/admin_chat/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
aiomysql>=0.2.0
aiohttp>=3.9.0
pydantic>=2.0.0
matplotlib>=3.7.0
plotly>=5.17.0
pillow>=10.0.0
kaleido>=0.2.1
# Vanna AI for enhanced SQL generation
vanna>=0.5.0
# MCP SDK (optional - install when available)
# mcp>=0.1.0
